+++ src/hex/hex_view.rs
@@ -1,4 +1,3 @@
-
 use {
     super::{
         byte::Byte,