+++ src/lib.rs
@@ -1,5 +1,3 @@
-// #![ allow( dead_code, unused_imports ) ]
-
 #[macro_use]
 extern crate crossbeam;
 #[macro_use]